requirements.txt
streamlit
openai
chromadb
python-dotenv
langchain-community
langchain-text-splitters
pillow
pypdf
docx2txt
